create lobby str utility
