+++ src/a20_lobby_db_toolbox/_utils/str_a20.py
@@ -1,0 +1,10 @@
+def lobbys_str() -> str:
+    return "lobbys"
+
+
+def lobby_id_str() -> str:
+    return "lobby_id"
+
+
+def lobby_mstr_dir_str() -> str:
+    return "lobby_mstr_dir"
