+++ src/a20_lobby_db_toolbox/test_/test_lobby_path_doc.py
@@ -5,10 +5,13 @@ from src.a02_finance_logic._utils.strs_a02 import (
     world_id_str,
 )
 from src.a06_bud_logic._utils.str_a06 import event_int_str
-from src.a20_lobby_db_toolbox.lobby_path import (
-    create_lobby_dir_path,
+from src.a20_lobby_db_toolbox._utils.str_a20 import (
+    lobbys_str,
     lobby_id_str,
     lobby_mstr_dir_str,
+)
+from src.a20_lobby_db_toolbox.lobby_path import (
+    create_lobby_dir_path,
     create_world_dir_path,
     create_fisc_mstr_dir_path,
 )
